backend/requirements.txt
Django==4.2.9
pillow==10.2.0
djangorestframework==3.14
djoser==2.2.2
django-filter==23.5
drf-extra-fields==3.7.0
psycopg2-binary==2.9.9
gunicorn==20.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
jsonschema==4.21.1
pytest-lazy-fixture==0.6.3